requirements.txt
supabase==2.18.1
storage3==0.12.2
requests==2.32.5
pandas==2.3.2
openpyxl==3.1.5
xlrd==2.0.2
PyYAML==6.0.2
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
tenacity==9.1.2
numpy==2.3.2